Separate server tick to its own file
